Namco System 2 ROZ is a device (so only gets created on games that have it, avoids a crash in old code in tilemap viewer)
